refactor(grid): rename helper module to helpers
Rename helper module to helpers for consistency.
