+++ src/goesdr/grid/helpers.py
